+++ src/main/java/nextstep/ladder/controller/LadderController.java
@@ -6,7 +6,7 @@ import nextstep.ladder.dto.ResultDto;
 import nextstep.ladder.dto.UniqueNames;
 import nextstep.ladder.factory.LadderFactory;
 import nextstep.ladder.generator.RandomNonConsecutiveFlagGenerator;
-import nextstep.ladder.interfaces.Ladder;
+import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.OutputView;
