+++ src/main/java/nextstep/ladder/domain/LadderImpl.java
@@ -2,9 +2,6 @@ package nextstep.ladder.domain;
 
 import java.util.List;
 import java.util.Objects;
-import nextstep.ladder.interfaces.Ladder;
-import nextstep.ladder.interfaces.Lines;
-import nextstep.ladder.interfaces.Position;
 
 public class LadderImpl implements Ladder {
