+++ src/main/java/nextstep/ladder/factory/LadderDataConverter.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import nextstep.ladder.generator.GeneratorWidthSize;
 import nextstep.ladder.generator.NonConsecutiveFlagGenerator;
-import nextstep.ladder.interfaces.Lines;
+import nextstep.ladder.domain.Lines;
 
 class LadderDataConverter {
     static List<Lines> convert(int height, int width, NonConsecutiveFlagGenerator generator) {
