+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -7,8 +7,6 @@ import java.util.Collections;
 import java.util.List;
 import nextstep.ladder.factory.LadderFactory;
 import nextstep.ladder.factory.LinesFactory;
-import nextstep.ladder.interfaces.Ladder;
-import nextstep.ladder.interfaces.Position;
 import org.junit.jupiter.api.Test;
 
 public class LadderTest {
