+++ src/test/java/nextstep/ladder/domain/LineTest.java
@@ -3,8 +3,6 @@ package nextstep.ladder.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import nextstep.ladder.interfaces.Line;
-import nextstep.ladder.interfaces.Position;
 import org.junit.jupiter.api.Test;
 
 public class LineTest {
